Move docker_ module_utils into subpackage.
