+++ lib/ansible/module_utils/docker/docker_common.py
