+++ lib/ansible/module_utils/docker/docker_swarm.py
@@ -7,11 +7,11 @@ import json
 try:
     from docker.errors import APIError
 except ImportError:
-    # missing docker-py handled in ansible.module_utils.docker_common
+    # missing docker-py handled in ansible.module_utils.docker.docker_common
     pass
 
 from ansible.module_utils._text import to_native
-from ansible.module_utils.docker_common import AnsibleDockerClient
+from ansible.module_utils.docker.docker_common import AnsibleDockerClient
 
 
 class AnsibleDockerSwarmClient(AnsibleDockerClient):
