+++ lib/ansible/modules/cloud/docker/docker_host_facts.py
@@ -186,17 +186,17 @@ docker_disk_usage:
 
 '''
 
-from ansible.module_utils.docker_common import AnsibleDockerClient, DockerBaseClass
+from ansible.module_utils.docker.docker_common import AnsibleDockerClient, DockerBaseClass
 from ansible.module_utils._text import to_native
 
 try:
     from docker.errors import APIError, NotFound
 except ImportError:
-    # missing docker-py handled in ansible.module_utils.docker_common
+    # missing docker-py handled in ansible.module_utils.docker.docker_common
     pass
 
 try:
-    from ansible.module_utils.docker_common import docker_version, clean_dict_booleans_for_docker_api
+    from ansible.module_utils.docker.docker_common import docker_version, clean_dict_booleans_for_docker_api
 except Exception as dummy:
     # missing docker-py handled in ansible.module_utils.docker
     pass
