+++ lib/ansible/modules/cloud/docker/docker_network_facts.py
@@ -110,7 +110,7 @@ docker_network:
     }'
 '''
 
-from ansible.module_utils.docker_common import AnsibleDockerClient
+from ansible.module_utils.docker.docker_common import AnsibleDockerClient
 
 
 def main():
