+++ lib/ansible/modules/cloud/docker/docker_node_facts.py
@@ -77,12 +77,12 @@ node_facts:
 
 from ansible.module_utils._text import to_native
 
-from ansible.module_utils.docker_swarm import AnsibleDockerSwarmClient
+from ansible.module_utils.docker.docker_swarm import AnsibleDockerSwarmClient
 
 try:
     from docker.errors import APIError, NotFound
 except ImportError:
-    # missing docker-py handled in ansible.module_utils.docker_common
+    # missing docker-py handled in ansible.module_utils.docker.docker_common
     pass
