+++ lib/ansible/modules/cloud/docker/docker_prune.py
@@ -173,10 +173,10 @@ builder_cache_space_reclaimed:
 
 from distutils.version import LooseVersion
 
-from ansible.module_utils.docker_common import AnsibleDockerClient
+from ansible.module_utils.docker.docker_common import AnsibleDockerClient
 
 try:
-    from ansible.module_utils.docker_common import docker_version, clean_dict_booleans_for_docker_api
+    from ansible.module_utils.docker.docker_common import docker_version, clean_dict_booleans_for_docker_api
 except Exception as dummy:
     # missing docker-py handled in ansible.module_utils.docker
     pass
