+++ lib/ansible/modules/cloud/docker/docker_volume.py
@@ -131,10 +131,10 @@ facts:
 try:
     from docker.errors import APIError
 except ImportError:
-    # missing docker-py handled in ansible.module_utils.docker_common
+    # missing docker-py handled in ansible.module_utils.docker.docker_common
     pass
 
-from ansible.module_utils.docker_common import (
+from ansible.module_utils.docker.docker_common import (
     DockerBaseClass,
     AnsibleDockerClient,
     DifferenceTracker,
