+++ test/units/module_utils/docker/test_docker_common.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ansible.module_utils.docker_common import (
+from ansible.module_utils.docker.docker_common import (
     compare_dict_allow_more_present,
     compare_generic,
 )
